TASK#15: Refactoring 1. Renames reducer.js module to root-reducer.js 2. Creates AxiosInstance inside store module and exports
