+++ project/src/components/header/user/user.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link, useHistory } from 'react-router-dom';
 import { AppRoute } from '../../../constants';
-import { State } from '../../../store/reducer';
+import { State } from '../../../store/root-reducer';
 import { connect, ConnectedProps } from 'react-redux';
 import { AsyncDispatch } from '../../../types/action';
 import { logout } from '../../../store/user/user-thunks';
