+++ project/src/index.tsx
@@ -3,31 +3,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './app';
 import { Provider } from 'react-redux';
-import { applyMiddleware, createStore } from '@reduxjs/toolkit';
-import { rootReducer } from './store/reducer';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import thunk from 'redux-thunk';
-import { createAPI } from './services/api';
-import { AxiosError, AxiosResponse } from 'axios';
-import { AuthStatus, HttpCode } from './constants';
-import { setAuthStatus } from './store/user/user-actions';
+import { store } from './store/store';
 
-export const api = createAPI();
-
-const store = createStore(rootReducer, composeWithDevTools(applyMiddleware(thunk.withExtraArgument(api))));
-
-api.interceptors.response.use(
-  (response: AxiosResponse) => response,
-  (error: AxiosError) => {
-    const { response } = error;
-
-    if (response?.status === HttpCode.Unauthorized) {
-      return store.dispatch(setAuthStatus(AuthStatus.NoAuth));
-    }
-
-    return Promise.reject(error);
-  },
-);
 
 ReactDOM.render(
   <React.StrictMode>
