+++ project/src/store/data/data-thunks.test.ts
@@ -1,8 +1,8 @@
 import MockAdapter from 'axios-mock-adapter';
 import thunk, { ThunkDispatch } from 'redux-thunk';
 import { configureMockStore } from '@jedmao/redux-mock-store';
-import { createAPI } from '../../services/api';
-import { State } from '../reducer';
+import { api } from '../../store/store';
+import { State } from '../root-reducer';
 import { Action } from 'redux';
 import { BackendRoute } from '../../constants';
 import { getFakeServerFilm, getFakeServerFilms } from '../../utils/mock';
@@ -13,7 +13,6 @@ import { setMyListButtonDisabled } from '../film/film-actions';
 
 
 describe('Data thunks', () => {
-  const api = createAPI();
   const mockApi = new MockAdapter(api);
   const middleware = [thunk.withExtraArgument(api)];
   const mockStore = configureMockStore<
