+++ project/src/store/user/user-thunks.ts
@@ -8,29 +8,47 @@ import { setAuthStatus, setUserData } from './user-actions';
 
 const login = (signIn: SignInForm): AsyncAction =>
   async (dispatch, _getState, api): Promise<void> => {
-    const { data } = await api.post<ServerUser>(BackendRoute.Login, signIn);
-    if (data) {
-      dispatch(setAuthStatus(AuthStatus.Auth));
-      dispatch(setUserData(adaptUserToClient(data)));
-      setToken(data.token);
+    try {
+      const { data } = await api.post<ServerUser>(BackendRoute.Login, signIn);
+
+      if (data) {
+        dispatch(setAuthStatus(AuthStatus.Auth));
+        dispatch(setUserData(adaptUserToClient(data)));
+        setToken(data.token);
+      }
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error);
     }
+
   };
 
 const logout = (): AsyncAction =>
   async (dispatch, _getState, api): Promise<void> => {
-    await api.delete(BackendRoute.Logout);
-    dispatch(setAuthStatus(AuthStatus.NoAuth));
-    dispatch(setUserData(null));
-    dropToken();
+    try {
+      await api.delete(BackendRoute.Logout);
+      dispatch(setAuthStatus(AuthStatus.NoAuth));
+      dispatch(setUserData(null));
+      dropToken();
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error);
+    }
+
   };
 
 const checkAuthStatus = (): AsyncAction =>
   async (dispatch, _getState, api): Promise<void> => {
-    const { data } = await api.get<ServerUser>(BackendRoute.Login);
+    try {
+      const { data } = await api.get<ServerUser>(BackendRoute.Login);
 
-    if (data) {
-      dispatch(setAuthStatus(AuthStatus.Auth));
-      dispatch(setUserData(adaptUserToClient(data)));
+      if (data) {
+        dispatch(setAuthStatus(AuthStatus.Auth));
+        dispatch(setUserData(adaptUserToClient(data)));
+      }
+    } catch (error) {
+      // eslint-disable-next-line no-console
+      console.error(error);
     }
   };
