+++ project/src/types/action.ts
@@ -3,7 +3,7 @@ import { AxiosInstance } from 'axios';
 import { AppActions } from '../store/app/app-actions';
 import { DataActions } from '../store/data/data-actions';
 import { FilmActions } from '../store/film/film-actions';
-import { State } from '../store/reducer';
+import { State } from '../store/root-reducer';
 import { unknownAction } from '../store/unknown';
 import { UserActions } from '../store/user/user-actions';
